feat: add session store for user and move user context
Improves project structure by relocating the UserContext to a
dedicated user-state module.

Introduces a user session store to encapsulate user session
management logic. This provides a centralized mechanism for
managing user authentication state.
