+++ src/lib/components/composites/project-components/CreateProjectDialog.svelte
@@ -12,7 +12,7 @@
     import { getContext, onMount } from "svelte";
     import { loadUsers } from "$lib/components/composites/input/loading-users";
     import AlertDialog from "$lib/components/composites/dialog/AlertDialog.svelte";
-    import { UserContextKey } from "$lib/global-context/userContext";
+    import { UserContextKey } from "$lib/user-state/userContext";
     import Alert from "../utils/Alert.svelte";
     import LoadingButton from "../button/LoadingButton.svelte";
     import { loadingWrapper } from "$lib/utils/common-helper";
